+++ packages/core/src/execution-engine/node-execution-context/utils/__tests__/request-helper-functions.test.ts
@@ -19,7 +19,6 @@ import type { ExecutionLifecycleHooks } from '@/execution-engine/execution-lifec
 import {
 	applyPaginationRequestData,
 	convertN8nRequestToAxios,
-	createFormDataObject,
 	httpRequest,
 	invokeAxios,
 	parseRequestObject,
@@ -27,6 +26,7 @@ import {
 	refreshOAuth2Token,
 	removeEmptyBody,
 } from '../request-helper-functions';
+import { createFormDataObject } from '../request-helpers/axios-utils';
 
 describe('Request Helper Functions', () => {
 	describe('proxyRequestToAxios', () => {